+++ zarrs_metadata/CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `ArrayMetadataV2` constructors
 - Implement `From<{&str,String}>` for `DataTypeMetadataV2`
 - Add `v2::array::codec::vlen_{array,bytes,utf8}` modules
+- Add support for Zarr V2 string fill values
 
 ### Changed
 - **Breaking**: Mark `GroupMetadataV3` and `ArrayMetadataV3` as non-exhaustive